requirements.txt
google-adk==1.14.0
google-cloud-alloydb-connector[pg8000]==1.6.0
sqlalchemy==2.0.41
pg8000==1.31.2
python-dotenv==1.1.0
